+++ kn/kmc/include/MpiClusters.h
@@ -6,18 +6,18 @@ namespace kn {
 
 class MpiClusters : public MpiIterator {
   public:
-    MpiClusters(long long int initial_number,
-                long long int increment_number,
-                long long int finial_number,
+    MpiClusters(unsigned long long int initial_number,
+                unsigned long long int increment_number,
+                unsigned long long int finial_number,
                 std::string solvent_element,
-                int smallest_cluster_criteria,
-                int solvent_bond_criteria);
+                size_t smallest_cluster_criteria,
+                size_t solvent_bond_criteria);
     virtual ~MpiClusters();
     void IterateToRun() override;
   private:
     std::string solvent_element_;
-    int smallest_cluster_criteria_{};
-    int solvent_bond_criteria_{};
+    size_t smallest_cluster_criteria_{};
+    size_t solvent_bond_criteria_{};
 };
 
 }// namespace kn
